chore(web-vis): extract wasm vector APIs into a wrapper with listener
